requirements.txt
autopep8==1.2.4
decorator==4.0.10
Django==1.10
django-cors-headers==2.0.1
django-extensions==1.7.4
django-haystack==2.6.0
django-model-utils==2.6
djangorestframework==3.4.7
elasticsearch==5.1.0
flake8==3.0.4
flasgger==0.5.13
geopy==1.11.0
ipython==5.1.0
ipython-genutils==0.1.0
jsonschema==2.5.1
pep8==1.7.0
psycopg2==2.6.2
gunicorn==19.6.0
whitenoise==3.2
python-decouple==3
dj-database-url==0.3.0
